refactor(tools): move vite-plus config under tools
