+++ docs/content/pt-BR/architecture/source-guide.md
@@ -117,7 +117,7 @@ A automação de repositórios prefere pacotes de comandos MoonBit sob `tools/mo
 do pacote (`moon run --target native tools/moon/cmd/<name> -- <args>`), compartilham a
 da cadeia de ferramentas que já constrói o compilador e são cobertos por suítes `tests/tooling/*.test.ts` que os
 via `moon run` e afirmam a saída esperada completa. As tarefas raiz as invocam com o ajudante `moonScript`
-em `config/vite-plus/task-commands.ts`, então cada consumidor mantém um nome de tarefa estável em vez de
+em `tools/config/vite-plus/task-commands.ts`, então cada consumidor mantém um nome de tarefa estável em vez de
 um comando inline.
 
 Bons candidatos ao MoonBit são pequenos, puros e com pouca dependência: análise sintática de argumentos, transformações de
