+++ tools/config/vite-plus/root-build-task-plugin.ts
