+++ tools/config/vite-plus/task-definition.ts
