+++ tools/config/vite-plus/task-shell.ts
