+++ tools/config/vite-plus/task-types.ts
