+++ tools/config/vite-plus/tasks/build.ts
