+++ tools/config/vite-plus/tasks/check.ts
