+++ tools/config/vite-plus/tasks/generation-cli.ts
