+++ tools/config/vite-plus/tasks/setup-dev.ts
