+++ tsconfig.node.json
@@ -16,5 +16,5 @@
     "composite": true,
     "tsBuildInfoFile": "./node_modules/.cache/tsconfig.node.tsbuildinfo"
   },
-  "include": ["tests/tooling/**/*.ts", "config/vite-plus/**/*.ts", "vite.config.ts"]
+  "include": ["tests/tooling/**/*.ts", "tools/config/vite-plus/**/*.ts", "vite.config.ts"]
 }
