+++ vite.config.ts
@@ -1,8 +1,8 @@
 import type { UserConfig } from "vite-plus";
 import { defineConfig } from "vite-plus";
-import { floatingPromiseTestPatterns } from "./config/vite-plus/task-inputs.ts";
-import { taskCatalog } from "./config/vite-plus/task-groups.ts";
-import { rootBuildTaskPlugin } from "./config/vite-plus/task-helpers.ts";
+import { floatingPromiseTestPatterns } from "./tools/config/vite-plus/task-inputs.ts";
+import { taskCatalog } from "./tools/config/vite-plus/task-groups.ts";
+import { rootBuildTaskPlugin } from "./tools/config/vite-plus/task-helpers.ts";
 
 const localGeneratedIgnorePatterns = [
   ".cache/**",
@@ -52,7 +52,7 @@ const formatterSensitiveContentIgnorePatterns = [
  * Root Vite+ configuration.
  *
  * The root config intentionally stays small: task helpers, package inputs, and
- * the task catalog live under `config/vite-plus/` where they can carry richer
+ * the task catalog live under `tools/config/vite-plus/` where they can carry richer
  * documentation and tighter type boundaries. This file should remain the place
  * that wires Vite+, repository lint/format policy, and the assembled task map
  * together.
